fix bug of index-out-of-bound

+++ xpath/params.h
@@ -80,11 +80,11 @@ extern unsigned int xpath_tlb_dre_capacity_bit;
 extern unsigned int xpath_tlb_dre_quantized_bit;
 
 /* CLOVE initial weight */
-extern unsigned int xpath_clove_init_weight;
+extern u16 xpath_clove_init_weight;
 /* The time gap between weight reduction for CLOVE */
 extern unsigned int xpath_clove_reduce_gap;
 /* CLOVE weight reset threshold */
-extern unsigned int xpath_clove_weight_reset_threshold;
+extern u16 xpath_clove_weight_reset_threshold;
 
 struct xpath_param
 {
